web-animations/animation-player/animation-player-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player currentTime attribute test</title>
<meta name="assert" content="Animation player currentTime attribute returns the effective current time of this player. Setting this attribute follows the procedure defined in section 'Performing a seek'">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-currentTime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function newAnimation(animationTarget) {
    animationTarget.style.width = '300px';
    return new Animation(animationTarget, [
	    {width: '10px', offset: 0},
	    {width: '100px', offset: 1/2},
	    {width: '200px', offset: 1}], 1000);
}


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
	
    // FIXME Now (2014, June) it's according the spec but may be changed in future.
    // Please check the specification if this test is still valid
    assert_not_equals(player.currentTime, null,
        'The value of AnimationPlayer.currentTime attribute should not be null');
}, 'Test value of AnimationPlayer.currentTime attribute should not be null');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
    player.pause();

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute when player is paused');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
    player.playbackRate = 0;

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is zero');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
    player.playbackRate = -1;

    var newCurrentTime = -100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is < 0 and seek time < 0');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    var newCurrentTime = player.source.endTime;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is > 0 and seek time equals to the source content end');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    var newCurrentTime = player.source.endTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player playback rate is > 0 and seek time > source content end');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));

    var seekTime = player.currentTime + 100;
    var expectedTime = player.timeline.currentTime - seekTime;
    player.currentTime = seekTime;

    assert_equals(player.currentTime, expectedTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
    assert_false(player.finished, 'Player\'s finished flag should be set to false');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player is not paused and playback rate is not zero');


test(function() {
    var player = document.timeline.play(newAnimation(createDiv(this)));
    player.playbackRate = 0.4;

    var seekTime = player.currentTime + 100;
    var expectedTime = player.timeline.currentTime - seekTime / player.playbackRate;
    player.currentTime = seekTime;

    assert_equals(player.currentTime, expectedTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
    assert_false(player.finished, 'Player\'s finished flag should be set to false');
}, 'Test setting value of AnimationPlayer.currentTime attribute ' +
    'when player is not paused and playback rate is not zero and not 1');


test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = 0;

    assert_equals(window.getComputedStyle(animation.target).width, 10,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is zero');


test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = 500;

    assert_equals(window.getComputedStyle(animation.target).width, 100,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time is > 0 and < end time');


test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = player.source.endTime;

    assert_equals(window.getComputedStyle(animation.target).width, 200,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute when current time equals end time');
</script>
</body>
